+++ packages/synthetics-sdk-broken-links/src/navigation_func.ts
@@ -17,11 +17,8 @@ import { pLimit } from 'plimit-lit';
 import {
   BrokenLinksResultV1_BrokenLinkCheckerOptions,
   BrokenLinksResultV1_SyntheticLinkResult,
-  GenericResultV1,
-  getRuntimeMetadata,
   ResponseStatusCode,
   ResponseStatusCode_StatusClass,
-  SyntheticResult,
 } from '@google-cloud/synthetics-sdk-api';
 import {
   checkStatusPassing,
@@ -303,25 +300,3 @@ async function fetchLink(
   const linkEndTime = new Date().toISOString();
   return { responseOrError, linkStartTime, linkEndTime };
 }
-
-// TODO move this to utils
-const getGenericError = (genericErrorMessage: string): GenericResultV1 => ({
-  ok: false,
-  generic_error: {
-    error_type: 'Error',
-    error_message: genericErrorMessage,
-    function_name: '',
-    file_path: '',
-    line: 0,
-  },
-});
-
-export const getGenericSyntheticResult = (
-  startTime: string,
-  genericErrorMessage: string
-): SyntheticResult => ({
-  synthetic_generic_result_v1: getGenericError(genericErrorMessage),
-  runtime_metadata: getRuntimeMetadata(),
-  start_time: startTime,
-  end_time: new Date().toISOString(),
-});
