+++ packages/synthetics-sdk-broken-links/test/unit/link_utils.spec.ts
@@ -20,22 +20,15 @@ import {
   ResponseStatusCode,
   ResponseStatusCode_StatusClass,
 } from '@google-cloud/synthetics-sdk-api';
-import {
-  BrokenLinkCheckerOptions,
-  LinkOrder,
-  StatusClass,
-} from '../../src/broken_links';
+import { LinkOrder, StatusClass } from '../../src/broken_links';
 import {
   checkStatusPassing,
   createSyntheticResult,
   LinkIntermediate,
   shouldGoToBlankPage,
   shuffleAndTruncate,
-  validateInputOptions,
 } from '../../src/link_utils';
-import {
-  setDefaultOptions,
-} from '../../src/options_func';
+import { setDefaultOptions } from '../../src/options_func';
 
 describe('GCM Synthetics Broken Links Utilies', async () => {
   const status_value_200: ResponseStatusCode = { status_value: 200 };
